[IMP] *website_sale*: split WebsiteSale interaction into smaller parts
task-5089578
