+++ addons/website_sale/static/src/interactions/quick_reorder.js
@@ -141,14 +141,14 @@ export class QuickReorder extends Interaction {
         }));
 
         // Add the product to the cart and update the DOM.
-        const cart = document.getElementById('shop_cart');
+        const websiteSale = document.querySelector('.oe_website_sale');
         // `updateCartNavBar` regenerates the cart lines and `updateQuickReorderSidebar`
         // regenerates the quick reorder products, so we need to stop and start interactions to
         // make sure the regenerated reorder products and cart lines are properly handled.
-        this.services['public.interactions'].stopInteractions(cart);
+        this.services['public.interactions'].stopInteractions(websiteSale);
         wSaleUtils.updateCartNavBar(data);
         wSaleUtils.updateQuickReorderSidebar(data);
-        this.services['public.interactions'].startInteractions(cart);
+        this.services['public.interactions'].startInteractions(websiteSale);
 
         // Move the focus to the next quantity input.
         this._focusNextQuantityInput(currentButtonIndex);
