+++ addons/website_sale/static/src/website_builder/product_page_option_plugin.js
@@ -177,10 +177,10 @@ export class ProductPageImageLayoutAction extends WebsiteConfigAction {
     static id = "productPageImageLayout";
     static dependencies = [...super.dependencies, "customizeWebsite", "productPageOption"];
     isApplied({ editingElement: productDetailMainEl, value }) {
-        return productDetailMainEl.dataset.image_layout === value;
+        return productDetailMainEl.dataset.imageLayout === value;
     }
     getValue({ editingElement: productDetailMainEl }) {
-        return productDetailMainEl.dataset.image_layout;
+        return productDetailMainEl.dataset.imageLayout;
     }
     async apply({ value }) {
         return rpc("/shop/config/website", { product_page_image_layout: value });
