+++ addons/website_sale/static/tests/builder/product_page_option.test.js
@@ -27,7 +27,7 @@ test("Product page options", async () => {
                         o_wsale_product_page_opt_image_radius_none
                         o_wsale_product_page_opt_separators"
                 >
-                    <div id="product_detail_main" data-image_layout="carousel">
+                    <div id="product_detail_main" data-image-layout="carousel">
                         <div class="o_wsale_product_images" data-image-amount="2">
                             <div id="o-carousel-product">
                                 <div class="carousel-item h-100 text-center active o_colored_level" style="min-height: 693px;">
